+++ src/datastructure/fastvertexset.h
@@ -16,7 +16,7 @@ public:
     typedef container::const_iterator const_iterator;
     typedef std::vector<value_type>::size_type size_type;
 
-    explicit FastVertexSet(unsigned int capacity = 0U) {
+    explicit FastVertexSet(unsigned long long capacity = 0ULL) {
         setIndex.setDefaultValue(0U);
         if (capacity > 0U) {
             setIndex.resetAll(capacity);
@@ -65,7 +65,7 @@ public:
 
 private:
     std::vector<Vertex*> set;
-    FastPropertyMap<unsigned int> setIndex;
+    FastPropertyMap<unsigned long long> setIndex;
 };
 
 }
